🛡️ Sentinel: [HIGH] Secure unprotected API routes and refactor authentication logic
- Add `verifyApiAuth` utility for consistent session and role verification.
- Protect `/api/fftt/players` and `/api/teams/matches` routes.
- Refactor existing routes for consistency and correct initialization order.
- Update OpenAPI documentation.
- Add security journal entry.
