+++ src/app/api/teams/[teamId]/matches/route.ts
@@ -1,8 +1,8 @@
-import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-import { cookies } from "next/headers";
-import { initializeFirebaseAdmin, getFirestoreAdmin, adminAuth } from "@/lib/firebase-admin";
+import type { NextRequest } from "next/server";
+import { getFirestoreAdmin } from "@/lib/firebase-admin";
 import { getTeamMatches } from "@/lib/server/team-matches";
+import { verifyApiAuth } from "@/lib/auth/api-auth";
 
 export const runtime = "nodejs";
 
@@ -10,44 +10,21 @@ export async function GET(
   _request: NextRequest,
   { params }: { params: Promise<{ teamId: string }> }
 ) {
-  // Vérification d'authentification
-  const cookieStore = await cookies();
-  const sessionCookie = cookieStore.get("__session")?.value;
-  
-  if (!sessionCookie) {
-    return NextResponse.json(
-      { error: "Authentification requise" },
-      { status: 401 }
-    );
-  }
-
   try {
-    const decoded = await adminAuth.verifySessionCookie(sessionCookie, true);
-    if (!decoded.email_verified) {
+    // Vérification d'authentification
+    const { errorResponse } = await verifyApiAuth();
+    if (errorResponse) return errorResponse;
+
+    // Récupérer teamId depuis les paramètres de route
+    const { teamId } = await params;
+
+    if (!teamId || typeof teamId !== "string") {
       return NextResponse.json(
-        { error: "Email non vérifié" },
-        { status: 403 }
+        { error: "Team ID parameter is required" },
+        { status: 400 }
       );
     }
-  } catch {
-    return NextResponse.json(
-      { error: "Session invalide" },
-      { status: 401 }
-    );
-  }
 
-  // Récupérer teamId depuis les paramètres de route
-  const { teamId } = await params;
-
-  if (!teamId || typeof teamId !== "string") {
-    return NextResponse.json(
-      { error: "Team ID parameter is required" },
-      { status: 400 }
-    );
-  }
-
-  try {
-    await initializeFirebaseAdmin();
     const firestore = getFirestoreAdmin();
     const matches = await getTeamMatches(firestore, teamId);
 
@@ -74,5 +51,3 @@ export async function GET(
     );
   }
 }
-
-
