+++ src/app/api/teams/route.ts
@@ -1,48 +1,18 @@
 import { NextResponse } from "next/server";
-import { cookies } from "next/headers";
-import { initializeFirebaseAdmin, getFirestoreAdmin, adminAuth } from "@/lib/firebase-admin";
+import { getFirestoreAdmin } from "@/lib/firebase-admin";
 import { getTeams, TeamSummary } from "@/lib/server/team-matches";
-import { hasAnyRole, USER_ROLES, resolveRole } from "@/lib/auth/roles";
+import { USER_ROLES } from "@/lib/auth/roles";
+import { verifyApiAuth } from "@/lib/auth/api-auth";
 
 export const runtime = "nodejs";
 
 export async function GET() {
-  // Vérification d'authentification
-  const cookieStore = await cookies();
-  const sessionCookie = cookieStore.get("__session")?.value;
-  
-  if (!sessionCookie) {
-    return NextResponse.json(
-      { error: "Authentification requise" },
-      { status: 401 }
-    );
-  }
-
   try {
-    const decoded = await adminAuth.verifySessionCookie(sessionCookie, true);
-    if (!decoded.email_verified) {
-      return NextResponse.json(
-        { error: "Email non vérifié" },
-        { status: 403 }
-      );
-    }
+    // Vérification d'authentification et d'autorisation
+    // L'accès à la liste des équipes est restreint aux administrateurs et coachs
+    const { errorResponse } = await verifyApiAuth([USER_ROLES.ADMIN, USER_ROLES.COACH]);
+    if (errorResponse) return errorResponse;
 
-    // Vérifier que l'utilisateur est admin ou coach
-    const role = resolveRole(decoded.role as string | undefined);
-    if (!hasAnyRole(role, [USER_ROLES.ADMIN, USER_ROLES.COACH])) {
-      return NextResponse.json(
-        { error: "Accès refusé" },
-        { status: 403 }
-      );
-    }
-  } catch {
-    return NextResponse.json(
-      { error: "Session invalide" },
-      { status: 401 }
-    );
-  }
-  try {
-    await initializeFirebaseAdmin();
     const firestore = getFirestoreAdmin();
 
     const teams: TeamSummary[] = await getTeams(firestore);
@@ -57,24 +27,9 @@ export async function GET() {
       { status: 200 }
     );
   } catch (error) {
-    console.error("[app/api/teams] Firestore Error:", error);
+    console.error("[app/api/teams] Error:", error);
     const errorMessage =
       error instanceof Error ? error.message : "Unknown error";
-
-    if (
-      errorMessage.includes("credentials") ||
-      errorMessage.includes("Could not load the default credentials")
-    ) {
-      return NextResponse.json(
-        {
-          error: "Firebase Admin credentials not configured",
-          details:
-            "Pour le développement local, configurez les credentials Firebase Admin. Voir README.md pour plus d'informations.",
-          message: errorMessage,
-        },
-        { status: 500 }
-      );
-    }
 
     return NextResponse.json(
       {
@@ -85,5 +40,3 @@ export async function GET() {
     );
   }
 }
-
-
